Used mathf.pow

+++ simulation_parameters/Constants.cs
@@ -446,7 +446,7 @@ public static class Constants
     public const float AUTO_EVO_SUNLIGHT_ENERGY_AMOUNT = 100000;
     public const float AUTO_EVO_COMPOUND_ENERGY_AMOUNT = 1200;
     public const float AUTO_EVO_CHUNK_ENERGY_AMOUNT = 90000000;
-    public const double AUTO_EVO_CHUNK_AMOUNT_NERF = 0.01;
+    public const float AUTO_EVO_CHUNK_AMOUNT_NERF = 0.01f;
     public const int AUTO_EVO_MINIMUM_SPECIES_SIZE_BEFORE_SPLIT = 80;
     public const bool AUTO_EVO_ALLOW_SPECIES_SPLIT_ON_NO_MUTATION = true;
